- All Daos check the apiKey / organisation before deleting

+++ src/main/java/info/interactivesystems/gamificationengine/dao/GoalDAO.java
@@ -91,7 +91,10 @@ public class GoalDAO {
 	 */
 	public Goal deleteGoalByIdAndOrganisation(int id, Organisation organisation) {
 		Goal goal = getGoalByIdAndOrganisation(id, organisation);
-		em.remove(goal);
+		
+		if(goal!= null){
+			em.remove(goal);
+		}
 		return goal;
 	}
